pipeline/requirements.txt
psycopg2-binary>=2.9.3
azure-storage-blob>=12.13.0
